[Refactor] Notion Database query
Notion DB query refactoring process
