+++ WhatToEat?/WhatToEat?/Views/MainTabView.swift
@@ -8,23 +8,15 @@ import SwiftUI
 
 struct MainTabView: View {
     @AppStorage("isFirstLaunch") var isFirstLaunch: Bool = false
-    @EnvironmentObject var modelData: GlobalModelData
 
     var body: some View {
         TabView {
-            RestaurantListView(modelData: modelData)
+            RestaurantListView()
                 .tabItem { Label("맛집", systemImage: "house")}
-            RecommendView(modelData: modelData)
+            RecommendView()
                 .tabItem { Label("추천", systemImage: "square.text.square") }
-            FavoriteRestaurantView(modelData: modelData)
+            FavoriteRestaurantView()
                 .tabItem { Label("먹킷리스트", systemImage: "person")}
-        }
-        .task {
-            if !isFirstLaunch {
-                await modelData.queryEntireNotionDB()
-                modelData.localData = modelData.serverData
-                isFirstLaunch = true
-            }
         }
     }
 }
