+++ WhatToEatBeforeRefac.xcworkspace/xcuserdata/jangjisu.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "8D96D813-EBBC-43CB-A93A-7D9BF5E8538C"
+   type = "0"
+   version = "2.0">
+</Bucket>
